Fix Auto Puzzle League
Using GameEnd(), I can detect when in post-game pre-menu (36) and if in post-game menu (16). Adding checks for if the value is either of these, zetris will no longer enter the ResetGame() code if in either of these areas now. On top of that, as far as I know, when the value hits 36, connection to your opponent no longer matters so this shouldn't cause any edge cases where zetris gets softlocked and can't unsoftlock.